fix: handle invalid CreatedOrUpdateTime
